Core refactor (#88)
* Develop (#85)

* Update project

* Update project

* Update project

* Update project

* Core refactor

* Core refactor
