+++ app/build.gradle.kts
@@ -151,7 +151,7 @@ android {
 }
 
 dependencies {
-    implementation(project(":core"))
+    implementation(project(":core:common"))
     implementation(project(":core:domain"))
     implementation(project(":core:navigation"))
     implementation(project(":feature:details"))
