+++ app/src/main/kotlin/org/michaelbel/movies/MainActivity.kt
@@ -9,7 +9,7 @@ import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.core.view.WindowCompat
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import dagger.hilt.android.AndroidEntryPoint
-import org.michaelbel.movies.core.shortcuts.installShortcuts
+import org.michaelbel.movies.common.shortcuts.installShortcuts
 import org.michaelbel.movies.ui.SystemTheme
 import org.michaelbel.movies.ui.theme.MoviesTheme
