+++ core/common/src/main/AndroidManifest.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <application>
+
+        <provider
+            android:name="androidx.startup.InitializationProvider"
+            android:authorities="${applicationId}.androidx-startup"
+            android:exported="false"
+            tools:node="merge">
+
+            <meta-data
+                android:name="org.michaelbel.movies.common.crashlytics.FirebaseCrashlyticsInitializer"
+                android:value="androidx.startup"
+                tools:node="remove"/>
+
+            <meta-data
+                android:name="org.michaelbel.movies.common.timber.TimberInitializer"
+                android:value="androidx.startup"
+                tools:node="remove"/>
+
+        </provider>
+
+    </application>
+
+</manifest>
