+++ core/common/src/main/kotlin/org/michaelbel/movies/common/GoogleApiModule.kt
@@ -1,0 +1,16 @@
+package org.michaelbel.movies.common
+
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object GoogleApiModule {
+
+    /*@Provides
+    @Singleton
+    fun provideGoogleApiAvailability(): GoogleApiAvailability {
+        return GoogleApiAvailability.getInstance()
+    }*/
+}
