+++ core/common/src/main/kotlin/org/michaelbel/movies/common/Urls.kt
@@ -1,0 +1,10 @@
+package org.michaelbel.movies.common
+
+const val EMAIL = "michael-bel@outlook.com"
+const val TELEGRAM_URL = "https://t.me/michaelbel"
+const val PAYPAL_ME = "https://paypal.me/michaelbel"
+const val GITHUB_URL = "https://github.com/michaelbel/moviemade"
+const val ACCOUNT_WEB = "https://play.google.com/store/apps/developer?id=Michael+Bel"
+const val ACCOUNT_MARKET = "market://developer?id=Michael+Bel"
+const val APP_WEB = "https://play.google.com/store/apps/details?id=org.michaelbel.moviemade"
+const val APP_MARKET = "market://details?id=org.michaelbel.moviemade"
