+++ core/common/src/main/kotlin/org/michaelbel/movies/common/config/RemoteParams.kt
@@ -1,0 +1,5 @@
+package org.michaelbel.movies.common.config
+
+object RemoteParams {
+    const val PARAM_SETTINGS_ICON_VISIBLE = "param_settings_icon_visible"
+}
