+++ core/common/src/main/kotlin/org/michaelbel/movies/common/coroutines/CoroutinesDispatchersModule.kt
@@ -1,0 +1,29 @@
+package org.michaelbel.movies.common.coroutines
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+
+@Module
+@InstallIn(SingletonComponent::class)
+object CoroutinesDispatchersModule {
+
+    @Provides
+    @Dispatcher(MoviesDispatchers.Default)
+    fun provideDefaultDispatcher(): CoroutineDispatcher = Dispatchers.Default
+
+    @Provides
+    @Dispatcher(MoviesDispatchers.IO)
+    fun provideIoDispatcher(): CoroutineDispatcher = Dispatchers.IO
+
+    @Provides
+    @Dispatcher(MoviesDispatchers.Main)
+    fun provideMainDispatcher(): CoroutineDispatcher = Dispatchers.Main
+
+    @Provides
+    @Dispatcher(MoviesDispatchers.MainImmediate)
+    fun provideMainImmediateDispatcher(): CoroutineDispatcher = Dispatchers.Main.immediate
+}
