+++ core/common/src/main/kotlin/org/michaelbel/movies/common/notifications/NotificationsModule.kt
@@ -1,0 +1,25 @@
+package org.michaelbel.movies.common.notifications
+
+import android.app.NotificationManager
+import android.content.Context
+import androidx.core.content.ContextCompat
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object NotificationsModule {
+
+    @Provides
+    fun provideNotificationService(
+        @ApplicationContext context: Context
+    ): NotificationManager {
+        return ContextCompat.getSystemService(
+            context,
+            NotificationManager::class.java
+        ) as NotificationManager
+    }
+}
