+++ core/common/src/main/kotlin/org/michaelbel/movies/common/permissions/PermissionKtx.kt
@@ -1,0 +1,12 @@
+package org.michaelbel.movies.common.permissions
+
+import android.content.Context
+import android.content.pm.PackageManager
+import androidx.core.content.ContextCompat
+
+fun String.denied(context: Context): Boolean {
+    return ContextCompat.checkSelfPermission(
+        context,
+        this
+    ) == PackageManager.PERMISSION_DENIED
+}
