+++ core/common/src/main/kotlin/org/michaelbel/movies/common/timber/TimberInitializer.kt
@@ -1,0 +1,20 @@
+package org.michaelbel.movies.common.timber
+
+import android.content.Context
+import androidx.startup.Initializer
+import org.michaelbel.movies.common.BuildConfig
+import org.michaelbel.movies.common.crashlytics.CrashlyticsTree
+import org.michaelbel.movies.common.crashlytics.FirebaseCrashlyticsInitializer
+import timber.log.Timber
+
+@Suppress("unused")
+class TimberInitializer: Initializer<Unit> {
+
+    override fun create(context: Context) {
+        Timber.plant(if (BuildConfig.DEBUG) Timber.DebugTree() else CrashlyticsTree())
+    }
+
+    override fun dependencies(): List<Class<out Initializer<*>>> {
+        return listOf(FirebaseCrashlyticsInitializer::class.java)
+    }
+}
