+++ core/common/src/main/kotlin/org/michaelbel/movies/common/viewmodel/BaseViewModel.kt
@@ -1,0 +1,33 @@
+package org.michaelbel.movies.common.viewmodel
+
+import androidx.annotation.CallSuper
+import androidx.lifecycle.ViewModel
+import kotlin.coroutines.CoroutineContext
+import kotlinx.coroutines.CoroutineExceptionHandler
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.cancelChildren
+import timber.log.Timber
+
+open class BaseViewModel: ViewModel(), CoroutineScope {
+
+    private val scopeJob: Job = SupervisorJob()
+
+    private val errorHandler = CoroutineExceptionHandler { _, throwable ->
+        handleError(throwable)
+    }
+
+    override val coroutineContext: CoroutineContext = scopeJob + Dispatchers.Main + errorHandler
+
+    override fun onCleared() {
+        coroutineContext.cancelChildren()
+        super.onCleared()
+    }
+
+    @CallSuper
+    protected open fun handleError(throwable: Throwable) {
+        Timber.e(throwable)
+    }
+}
