+++ core/domain/build.gradle.kts
@@ -23,14 +23,17 @@ android {
     buildTypes {
         release {
             isMinifyEnabled = true
-            proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
+            proguardFiles(
+                getDefaultProguardFile("proguard-android-optimize.txt"),
+                "proguard-rules.pro"
+            )
         }
     }
 }
 
 dependencies {
-    implementation(project(":core"))
     implementation(project(":core:analytics"))
+    implementation(project(":core:common"))
     implementation(project(":core:entities"))
     implementation(project(":core:network"))
     implementationDataStoreDependencies()
