+++ core/domain/src/main/kotlin/org/michaelbel/movies/domain/interactor/impl/MovieInteractorImpl.kt
@@ -3,7 +3,8 @@ package org.michaelbel.movies.domain.interactor.impl
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
-import org.michaelbel.movies.core.coroutines.IoDispatcher
+import org.michaelbel.movies.common.coroutines.Dispatcher
+import org.michaelbel.movies.common.coroutines.MoviesDispatchers
 import org.michaelbel.movies.domain.interactor.MovieInteractor
 import org.michaelbel.movies.domain.repository.MovieRepository
 import org.michaelbel.movies.entities.Either
@@ -11,7 +12,7 @@ import org.michaelbel.movies.entities.MovieData
 import org.michaelbel.movies.entities.MovieDetailsData
 
 class MovieInteractorImpl @Inject constructor(
-    @IoDispatcher private val dispatcher: CoroutineDispatcher,
+    @Dispatcher(MoviesDispatchers.IO) private val dispatcher: CoroutineDispatcher,
     private val movieRepository: MovieRepository
 ): MovieInteractor {
