+++ core/navigation/build.gradle.kts
@@ -14,7 +14,10 @@ android {
     buildTypes {
         release {
             isMinifyEnabled = true
-            proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
+            proguardFiles(getDefaultProguardFile(
+                "proguard-android-optimize.txt"),
+                "proguard-rules.pro"
+            )
         }
     }
 }
