--- core/src/main/AndroidManifest.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <application>
-
-        <provider
-            android:name="androidx.startup.InitializationProvider"
-            android:authorities="${applicationId}.androidx-startup"
-            android:exported="false"
-            tools:node="merge">
-
-            <meta-data
-                android:name="org.michaelbel.movies.core.crashlytics.FirebaseCrashlyticsInitializer"
-                android:value="androidx.startup"
-                tools:node="remove"/>
-
-            <meta-data
-                android:name="org.michaelbel.movies.core.timber.TimberInitializer"
-                android:value="androidx.startup"
-                tools:node="remove"/>
-
-        </provider>
-
-    </application>
-
-</manifest>
