--- core/src/main/kotlin/org/michaelbel/movies/core/GoogleApiModule.kt
@@ -1,16 +1,0 @@
-package org.michaelbel.movies.core
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-object GoogleApiModule {
-
-    /*@Provides
-    @Singleton
-    fun provideGoogleApiAvailability(): GoogleApiAvailability {
-        return GoogleApiAvailability.getInstance()
-    }*/
-}
