--- core/src/main/kotlin/org/michaelbel/movies/core/config/RemoteParams.kt
@@ -1,5 +1,0 @@
-package org.michaelbel.movies.core.config
-
-object RemoteParams {
-    const val PARAM_SETTINGS_ICON_VISIBLE = "param_settings_icon_visible"
-}
