--- core/src/main/kotlin/org/michaelbel/movies/core/coroutines/CoroutinesDispatchersModule.kt
@@ -1,29 +1,0 @@
-package org.michaelbel.movies.core.coroutines
-
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-
-@Module
-@InstallIn(SingletonComponent::class)
-object CoroutinesDispatchersModule {
-
-    @Provides
-    @DefaultDispatcher
-    fun provideDefaultDispatcher(): CoroutineDispatcher = Dispatchers.Default
-
-    @Provides
-    @IoDispatcher
-    fun provideIoDispatcher(): CoroutineDispatcher = Dispatchers.IO
-
-    @Provides
-    @MainDispatcher
-    fun provideMainDispatcher(): CoroutineDispatcher = Dispatchers.Main
-
-    @Provides
-    @MainImmediateDispatcher
-    fun provideMainImmediateDispatcher(): CoroutineDispatcher = Dispatchers.Main.immediate
-}
