+++ feature/feed/build.gradle.kts
@@ -48,8 +48,8 @@ android {
 }
 
 dependencies {
-    implementation(project(":core"))
     implementation(project(":core:analytics"))
+    implementation(project(":core:common"))
     implementation(project(":core:navigation"))
     implementation(project(":core:domain"))
     implementationHiltDependencies()
