+++ feature/settings/build.gradle.kts
@@ -51,9 +51,9 @@ android {
 }
 
 dependencies {
-    implementation(project(":core"))
     implementation(project(":core:ads"))
     implementation(project(":core:analytics"))
+    implementation(project(":core:common"))
     implementation(project(":core:domain"))
     implementation(project(":core:navigation"))
     implementation(project(":core:ui"))
