+++ releaseNotes.txt
@@ -1,0 +1,107 @@
+* Merge remote-tracking branch 'origin/develop' into develop
+* Features refactor (#87)
+* Merge branch 'master' into develop
+* Develop (#85)
+* Merge branch 'master' into develop
+* Delete source-to-classes.tab.values.at
+* Delete source-to-classes.tab.len
+* Delete source-to-classes.tab.keystream.len
+* Delete source-to-classes.tab.keystream
+* Delete source-to-classes.tab
+* Delete proto.tab_i.len
+* Delete proto.tab_i
+* Delete proto.tab.values.at
+* Delete proto.tab.len
+* Delete proto.tab.keystream.len
+* Delete proto.tab.keystream
+* Delete proto.tab
+* Delete internal-name-to-source.tab_i
+* Delete internal-name-to-source.tab.values.at
+* Delete source-to-classes.tab_i.len
+* Delete source-to-classes.tab_i
+* Delete internal-name-to-source.tab.len
+* Delete internal-name-to-source.tab.keystream.len
+* Delete internal-name-to-source.tab.keystream
+* Delete internal-name-to-source.tab
+* Delete internal-name-to-source.tab_i.len
+* Delete constants.tab_i.len
+* Delete constants.tab_i
+* Delete constants.tab.values.at
+* Delete constants.tab.len
+* Delete constants.tab.keystream.len
+* Delete constants.tab.keystream
+* Delete constants.tab
+* Delete class-fq-name-to-source.tab_i.len
+* Delete class-fq-name-to-source.tab_i
+* Delete class-fq-name-to-source.tab.values.at
+* Delete class-fq-name-to-source.tab.len
+* Delete class-fq-name-to-source.tab.keystream.len
+* Delete class-fq-name-to-source.tab.keystream
+* Delete class-fq-name-to-source.tab
+* Delete class-attributes.tab_i.len
+* Delete class-attributes.tab_i
+* Delete class-attributes.tab.values.at
+* Delete lookups.tab_i
+* Delete lookups.tab.values.at
+* Delete lookups.tab.len
+* Delete lookups.tab.keystream.len
+* Delete lookups.tab_i.len
+* Delete lookups.tab.keystream
+* Delete lookups.tab
+* Delete id-to-file.tab_i.len
+* Delete id-to-file.tab_i
+* Delete class-attributes.tab.len
+* Delete class-attributes.tab.keystream.len
+* Delete class-attributes.tab.keystream
+* Delete class-attributes.tab
+* Delete lint.xml
+* Delete .gitignore
+* Delete id-to-file.tab.values.at
+* Delete id-to-file.tab.keystream.len
+* Delete id-to-file.tab.keystream
+* Delete id-to-file.tab.len
+* Delete id-to-file.tab
+* Delete file-to-id.tab_i.len
+* Delete file-to-id.tab_i
+* Delete file-to-id.tab.values.at
+* Delete .gitignore
+* Update readme.md
+* Update project
+* Update readme.md
+* Update project
+* Update project
+* Update project
+* Update readme.md
+* Develop (#84)
+* Update readme.md
+* Update readme.md
+* Delete file-to-id.tab.len
+* Delete file-to-id.tab.keystream.len
+* Delete file-to-id.tab.keystream
+* Delete file-to-id.tab
+* Delete counters.tab
+* Update readme.md
+* Update project (#83)
+* Delete source-to-output.tab_i.len
+* Delete source-to-output.tab_i
+* Delete source-to-output.tab.values.at
+* Delete source-to-output.tab.len
+* Delete source-to-output.tab.keystream.len
+* Delete source-to-output.tab.keystream
+* Develop (#81)
+* Update distribution_app.yml
+* Update distribution_app.yml
+* Develop (#80)
+* Delete source-to-output.tab
+* Delete build-history.bin
+* Delete last-build.bin
+* Update readme.md
+* Delete buildSrc.kotlin_module
+* Delete App.class
+* Delete buildSrcjar-classes.txt
+* Delete buildSrc.jar
+* Delete plugin-under-test-metadata.properties
+* Delete validation-report.txt
+* Delete source-roots.txt
+* Delete MANIFEST.MF
+* Delete releaseNotes.txt
